src/bot/requirements.txt
python-telegram-bot==13.12
cachetools==4.2.2
requests==2.26.0
SQLAlchemy==1.4.28
pydantic==1.8.2
python-dotenv==0.20.0
python_dateutil==2.8.2
